убрал папку include, разобрался в с двумя ворнинграми, с третьим хз
