+++ GuiApplication/objects/map_screen.h
@@ -8,6 +8,7 @@
 class MapScreen : public QWidget {
 public:
     MapScreen(QWidget *parent = nullptr);
+
 };
 
 #endif // MAPSCREEN_H
