--- include/settings.pri
@@ -1,5 +1,0 @@
-HEADERS += \
-    $$PWD/settings.h
-
-SOURCES += \
-    $$PWD/settings.cpp
